Add files via upload (#3)
File handling is very raw at the moment. Saves value of ribbon checkboxes in profile editor activity and loads them into memory in another activity. XML approach is to be added. The profile file is being written into external storage public so that developers can better see the output and not guess how it is.